+++ index.html
@@ -53,6 +53,6 @@
       <p class="footer-text">Copyright</p>
     </div>
   </footer>
-  <script src="index.js"></script>
+  <script src="index.js" type="module"></script>
 </body>
 </html>
