+++ modules/selectors.js
@@ -1,0 +1,10 @@
+export const form = document.querySelector('#form');
+export const listBooks = document.querySelector('#list-books');
+
+export const list = document.querySelector('.list');
+export const addNew = document.querySelector('.add-new');
+export const contact = document.querySelector('.contact-info');
+
+export const navList = document.querySelector('#nav-list');
+export const navAddNew = document.querySelector('#navAddNew');
+export const navContact = document.querySelector('#navContact');
